% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopProjF.R
\name{PopProjF}
\alias{PopProjF}
\title{Title}
\usage{
PopProjF(x, fed = 10)
}
\arguments{
\item{x}{Data.frame with parameters from cmartr::PopParam}

\item{fed}{Density of females per km}
}
\value{
Data.frame with projections for each parameter provided in x.
}
\description{
Projects populations through different scenarios across 
species range.
}
\examples{
\dontrun{
dfpop <- cmartr::PopParam(species = "Podocnemis unifilis", make_rds = FALSE)
library(plyr)
dff <- ddply(dfpop, .(akey), .fun = PopProjF)
}
}
